chore(umi-ui): improve code style (#3493)
* chore(umi-ui): improve code style

* chore: improve block directory structure

* fix(umi-ui): UMI_UI env don't work in task

* chore: move UMI_UI enable check to block/index.js

* fix: umi ui task env locales

* refact: extract socket handlers from server/index.ts

* code style
