+++ packages/umi-build-dev/src/plugins/commands/block/ui/client/index.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { IUiApi } from 'umi-types';
 
-import Context from './ui/UIApiContext';
-import BlocksViewer from './ui/index';
-import Icon from './ui/icon';
+import Context from './UIApiContext';
+import BlocksViewer from './BlocksViewer';
+import Icon from './icon';
 import zhCN from './locales/zh-CN';
 import enUS from './locales/en-US';
-import model, { initApiToGlobal, namespace } from './ui/model';
+import model, { initApiToGlobal, namespace } from './model';
 
 export default (api: IUiApi) => {
   initApiToGlobal(api);
