+++ packages/umi-build-dev/src/plugins/commands/block/ui/server/socketHandlers/org.umi.block.add.ts
@@ -1,0 +1,12 @@
+import { AddBlockParams } from '../../../data';
+
+export default async function({ blockService, success, payload }) {
+  const { url } = payload as AddBlockParams;
+  await blockService.run({ ...payload });
+  success({
+    data: {
+      message: `🎊 ${url} block is adding`,
+    },
+    success: true,
+  });
+}
