+++ packages/umi-build-dev/src/plugins/commands/block/ui/server/socketHandlers/org.umi.block.cancel.ts
@@ -1,0 +1,6 @@
+export default async function({ blockService, success }) {
+  success({
+    data: blockService.cancel(),
+    success: true,
+  });
+}
