+++ packages/umi-build-dev/src/plugins/commands/block/ui/server/socketHandlers/org.umi.block.checkExistRoute.ts
@@ -1,0 +1,9 @@
+import { AddBlockParams } from '../../../data';
+
+export default function({ payload, blockService, success }) {
+  const { path } = payload as AddBlockParams;
+  success({
+    exists: blockService.routeExists(path),
+    success: true,
+  });
+}
