+++ packages/umi-build-dev/src/plugins/commands/block/ui/server/socketHandlers/org.umi.block.clear.ts
@@ -1,0 +1,9 @@
+import clearGitCache from '../../../clearGitCache';
+
+export default function({ payload, api, success }) {
+  const info = clearGitCache(payload, api);
+  success({
+    data: info,
+    success: true,
+  });
+}
