+++ packages/umi-build-dev/src/plugins/commands/block/ui/server/socketHandlers/org.umi.block.get-adding-block-url.ts
@@ -1,0 +1,6 @@
+export default function({ blockService, success }) {
+  success({
+    data: blockService.getBlockUrl(),
+    success: true,
+  });
+}
