+++ packages/umi-build-dev/src/plugins/commands/block/ui/server/socketHandlers/org.umi.block.resource.ts
@@ -1,0 +1,6 @@
+export default function({ success, resources }) {
+  success({
+    data: resources,
+    success: true,
+  });
+}
