+++ packages/umi-build-dev/src/plugins/commands/block/ui/server/socketHandlers/org.umi.block.retry.ts
@@ -1,0 +1,6 @@
+export default async function({ blockService, success, payload }) {
+  success({
+    data: blockService.retry({ ...payload }),
+    success: true,
+  });
+}
