+++ packages/umi-ui-tasks/src/locales/zh-CN.js
@@ -41,16 +41,15 @@ export default {
   'org.umi.ui.tasks.envs.BROWSER.desc': '自动打开浏览器。',
   'org.umi.ui.tasks.envs.clear': '日志清屏',
   'org.umi.ui.tasks.envs.clear.desc': '启动日志自动清除。',
-  'org.umi.ui.tasks.envs.tsCheck': 'TS 类型检查',
-  'org.umi.ui.tasks.envs.tsCheck.desc': 'TS 类型定义检查（适用于 TS 项目）。',
+  'org.umi.ui.tasks.envs.tsCheck': 'TypeScript 类型检查',
+  'org.umi.ui.tasks.envs.tsCheck.desc': 'TypeScript 类型定义检查（适用于 TypeScript 项目）。',
   'org.umi.ui.tasks.envs.umiUI': 'UmiUI',
-  'org.umi.ui.tasks.envs.umiUI.desc': '打开 umi ui。',
+  'org.umi.ui.tasks.envs.umiUI.desc': '打开 Umi UI。',
   // build
   'org.umi.ui.tasks.build.buildError': '执行构建失败',
   'org.umi.ui.tasks.build.cancelError': '取消构建失败',
   'org.umi.ui.tasks.build.start': '构建',
   'org.umi.ui.tasks.build.cancel': '停止',
-  'org.umi.ui.tasks.envs': '环境变量',
   'org.umi.ui.tasks.envs.compress': 'JS 压缩',
   'org.umi.ui.tasks.envs.compress.desc': '压缩 JS 文件。',
   'org.umi.ui.tasks.envs.cssCompress': 'CSS 压缩',
